세미나 5주차 실습 과제 완료

+++ src/main/java/com/project/likelion13thbe/domain/product/dto/response/ProductResDTO.java
@@ -14,7 +14,7 @@ public class ProductResDTO {
             Integer price,
             String image,
             String description,
-            Integer rating
+            double rating
     ) {}
 
     @Builder
